jwt: Fix review comments
char[] is used for password field to follow the best practices
in holding password values in java. This way value of the field
is not visible in memory dumps. This change needs to be made in
many other places but for now I'm making this change in the new
src added for secretProvider impl.